no shared

+++ src/matrix.cpp
@@ -11,7 +11,7 @@
 
 matrix_base::matrix_base(size_t rows, size_t cols) {
 	    DEBUGMEM("matrix: allocating double: %lu * %lu \n",rows,cols);
-		data = new double[rows*cols] ;
+		data = new float[rows*cols] ;
 		this->rows = rows;
 		this->cols = cols;
 		DEBUGMEM("matrix: allocated double: %lu * %lu in %p\n",rows,cols,data);
